+++ finagle-zipkin-scribe/src/main/scala/com/twitter/finagle/zipkin/thrift/ZipkinTracer.scala
@@ -1,8 +1,10 @@
 package com.twitter.finagle.zipkin.thrift
 
-import com.twitter.finagle.stats.{NullStatsReceiver, StatsReceiver}
+import com.twitter.finagle.stats.NullStatsReceiver
+import com.twitter.finagle.stats.StatsReceiver
 import com.twitter.finagle.tracing.Tracer
-import com.twitter.finagle.zipkin.{core, host => Host}
+import com.twitter.finagle.zipkin.core
+import com.twitter.finagle.zipkin.{host => Host}
 
 object ZipkinTracer {
 
@@ -19,7 +21,7 @@ object ZipkinTracer {
     scribeHost: String = Host().getHostName,
     scribePort: Int = Host().getPort,
     statsReceiver: StatsReceiver = NullStatsReceiver,
-    sampleRate: Float = core.Sampler.DefaultSampleRate
+    sampleRate: Float = core.DefaultSampler.sampleRate
   ): Tracer = mk(scribeHost, scribePort, statsReceiver, sampleRate)
 
   /**
@@ -32,7 +34,7 @@ object ZipkinTracer {
     host: String = Host().getHostName,
     port: Int = Host().getPort,
     statsReceiver: StatsReceiver = NullStatsReceiver,
-    sampleRate: Float = core.Sampler.DefaultSampleRate
+    sampleRate: Float = core.DefaultSampler.sampleRate
   ): Tracer =
     new ZipkinTracer(
       ScribeRawZipkinTracer(scribeHost = host, scribePort = port, statsReceiver = statsReceiver),
@@ -45,14 +47,14 @@ object ZipkinTracer {
    */
   @deprecated("Use mk() instead", "6.1.0")
   def apply(sr: StatsReceiver): Tracer =
-    mk(Host().getHostName, Host().getPort, sr, core.Sampler.DefaultSampleRate)
+    mk(Host().getHostName, Host().getPort, sr, core.DefaultSampler.sampleRate)
 
   /**
    * Util method since named parameters can't be called from Java
    * @param statsReceiver stats receiver to send successes/failures to
    */
   def mk(statsReceiver: StatsReceiver): Tracer =
-    mk(Host().getHostName, Host().getPort, statsReceiver, core.Sampler.DefaultSampleRate)
+    mk(Host().getHostName, Host().getPort, statsReceiver, core.DefaultSampler.sampleRate)
 }
 
 /**